ml-research/requirements.txt
tensorflow-cpu==2.15.0; python_version < "3.12"
tensorflow-cpu==2.21.0; python_version >= "3.12"
numpy==1.26.4; python_version < "3.13"
numpy==2.1.3; python_version >= "3.13"
pandas==2.2.2
scikit-learn==1.5.1
pytest==8.3.2
pyspark==3.5.1
